+++ database/migrations/2017_01_17_121428_make_param_varchar_on_permissions.php
@@ -1,0 +1,32 @@
+<?php
+
+use Carbon\Carbon;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class MakeParamVarcharOnPermissions extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('permissions', function($table) {
+            $table->string('param')->nullable()->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('permissions', function($table) {
+            $table->unsignedInteger('param')->nullable()->change();
+        });
+    }
+}